### 2025-11-21 #9493 @editorialbot
[PRE REVIEW]: PipeOptz: A Framework for Pipeline Optimization
**Submitting author:** @romsson (<a href="http://orcid.org/0000-0003-1447-6926">Romain Vuillemot</a>)
**Repository:** https://github.com/centralelyon/pipeoptz/
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.5
