### 2025-12-22 82bbfe3
phrasing and ref fixing
