### 2026-01-12 0dd4781
AI usage disclosure + Research Impact Statement + Software Design + State of the field
