+++ paper.bib
@@ -166,3 +166,56 @@ journal = {Journal of Open Source Software}
   url   = {https://www.prefect.io/},
   note  = {Accessed: 2025-12-21}
 }
+
+@article{Kim2023_BayesO,
+  doi = {10.21105/joss.05320},
+  url = {https://doi.org/10.21105/joss.05320},
+  year = {2023},
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {90},
+  pages = {5320},
+  author = {Kim, Jungtaek and Choi, Seungjin},
+  title = {BayesO: A Bayesian optimization framework in Python},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Hofmann2023_Linopy,
+  doi = {10.21105/joss.04823},
+  url = {https://doi.org/10.21105/joss.04823},
+  year = {2023},
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {84},
+  pages = {4823},
+  author = {Hofmann, Fabian},
+  title = {Linopy: Linear optimization with n-dimensional labeled variables},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Pecnik2021_NiaAML,
+  doi = {10.21105/joss.02949},
+  url = {https://doi.org/10.21105/joss.02949},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {61},
+  pages = {2949},
+  author = {Pe{\v{c}}nik, Luka and Fister, Iztok},
+  title = {NiaAML: AutoML framework based on stochastic population-based nature-inspired algorithms},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Jimenez2017_pyGPGO,
+  doi = {10.21105/joss.00431},
+  url = {https://doi.org/10.21105/joss.00431},
+  year = {2017},
+  publisher = {The Open Journal},
+  volume = {2},
+  number = {19},
+  pages = {431},
+  author = {Jim{\'e}nez, Jos{\'e} and Ginebra, Josep},
+  title = {pyGPGO: Bayesian Optimization for Python},
+  journal = {Journal of Open Source Software}
+}
+
