### 2026-01-13 #9493 @pengov
Hi @danielskatz, thanks for the update.

We have updated the manuscript to match the new required JOSS format. The paper now includes the additional required sections Software Design, Research Impact Statement, and an AI Usage Disclosure. We also expanded the "State of the field" section to better position PipeOptz relative to closely related JOSS software.

We did not contact any of the authors of these papers suggested by @editorialbot ; we only referenced some of them in the manuscript as prior work.
