+++ paper.bib
@@ -158,13 +158,15 @@ journal = {Journal of Open Source Software}
 @misc{airflow,
   title = {Apache Airflow},
   url   = {https://airflow.apache.org/},
-  note  = {Accessed: 2025-12-21}
+  note  = {Accessed: 2025-12-21},
+  urldate = {2025-12-21}
 }
 
 @misc{prefect,
   title = {Prefect},
   url   = {https://www.prefect.io/},
-  note  = {Accessed: 2025-12-21}
+  note  = {Accessed: 2025-12-21},
+  urldate = {2025-12-21}
 }
 
 @article{Kim2023_BayesO,
