### 2026-01-21 #9493 @gflofst
> good morning, [@gflofst](https://github.com/gflofst)! You have reviewed a paper like this for us before, would you be interested in reviewing this one?

Yes! JOSS is my favorite.
