### 2026-01-21 #9871 @erik-whiting
Hello @romsson, thank you again for submitting to JOSS. This GitHub issue is where we will coordinate the review of your paper. The reviewers will generate checklists and then open issues on your repository as they go through the review process.

Thank you so much @jqwang2373 and @gflofst for agreeing to review this paper! [Here is a link to the review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). Please note that the section titled "Using AI to generate code" is out of date, we _**do**_ have an AI use policy; here is a link to [the JOSS AI usage policy](https://joss.readthedocs.io/en/latest/submitting.html#ai-usage-policy).

While reviewing the work, please open an issue on [the repository under review](https://github.com/centralelyon/pipeoptz/) unless it is a minor thing which can be mentioned in a single, concise comment.

Please let me know if there are any questions, I look forward to working through this with you all!
