### 2026-02-27 #9871 @shrishar
@erik-whiting Just finished first round of reviewing. Here is my review. I will review again once the issues are addressed

1) Contribution and authorship: Has the submitting author (@romsson) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
_I see only one of the authors contributing the most in the repository. I dont see submitting author role in developing the code from the commit history. I might be missing something here. May be they have contributed in a different way but its not very clear from commit history_

2)  Collaborative effort: Does the commit history show contributions from multiple developers and evidence of iterative refinement through community feedback?
_Only one person has contributed the most out of three._

3) Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
_Please update this section. It talks about core principle, but I would like to see how these components interact with each other. I would like to the high level architecture design._

4) Research Impact Statement (required section): Does the paper include a "Research Impact Statement" section providing evidence of realized impact (publications, external use, integrations) or credible near-term significance (benchmarks, reproducible materials, community-readiness signals)? Is the evidence presented compelling?

_I dont see concreate results yet. The paper states that it is still ongoing. I would like to see concrete evidence that tool is optimizing/fine tuning as compared to traditional method through A/B testing if possible_

5) References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

_I believe they can add more citations if they can. I can check this off if it is not possible_

6) Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
_Add software architecture and results on how this tool is performing. Rest seems fine to me_
