### 2026-03-24 #9871 @gflofst
I’ve been on vacation. I’ll get started this week and done hopefully by next week some time.

Best,

Jay

From: Daniel S. Katz ***@***.***>
Sent: Monday, March 23, 2026 7:33 AM
Cc: Lofstead, Jay F ***@***.***>; Mention ***@***.***>
Subject: [EXTERNAL] Re: [openjournals/joss-reviews] [REVIEW]: PipeOptz: A Framework for Pipeline Optimization (Issue #9871)

[https://avatars.githubusercontent.com/u/2913845?s=20&v=4]danielskatz left a comment (openjournals/joss-reviews#9871)<https://github.com/openjournals/joss-reviews/issues/9871#issuecomment-4110631815>

👋 @romsson<https://github.com/romsson> - It appears the review process here is waiting for you.
