### 2026-04-03 9171c10
Define version only once
