### 2026-04-03 1a28081
Adding exception handling
