### 2026-04-05 da9771c
Create visualization.py as a new module based on Node
