### 2026-04-05 3936ab3
Adding the to_dot and to_image from pipeline
