### 2026-04-05 7fdda06
Updating the pipeline and init
