+++ src/pipeoptz/pipeline.py
@@ -11,6 +11,7 @@ from collections import deque
 from typing import Callable, Any, Union, Dict, List, \
                    Tuple, Deque, Iterator, Optional, Iterable
 from .node import Node, NodeIf, NodeFor, NodeWhile
+from .visualization import Visualizer
 
 
 
@@ -355,157 +356,14 @@ class Pipeline:
         Returns:
             the DOT string of the pipeline
         """
-        def escape_id(nid: str) -> str:
-            return f"{_prefix}{nid}"
-
-        dot_lines: List[str] = []
-        dot_lines.append("digraph Pipeline {" if _prefix == "" else "subgraph Pipeline {")
-        dot_lines.append('  rankdir=TB;')  # vertical layout
-        dot_lines.append('  node [fontsize=12 fontname="Helvetica"];')
-
-        last_node_id = self.static_order()[-1] if self.static_order() else None
-
-        for node_id, node in self.nodes.items():
-            full_id = escape_id(node_id).replace(" ", "_")
-            is_last = node_id == last_node_id
-
-            if isinstance(node, NodeIf):
-                if node.func.__module__ == "__main__":
-                    func_label = node.func.__name__
-                else:
-                    func_label = f"{node.func.__module__}.{node.func.__name__}"
-                if node.func.__name__ == "<lambda>":
-                    func_label = "lambda"
-                dot_lines.append(f'  subgraph cluster_{full_id} {{')
-                dot_lines.append('    style=dashed;')
-                if show_function:
-                    dot_lines.append(f'    "{full_id}" [shape=diamond, label=< <B>{node_id}</B> '+
-                                     f'<BR/><FONT POINT-SIZE=\"10\">{func_label}</FONT> >];')
-                else:
-                    dot_lines.append(f'    "{full_id}" [shape=diamond, \
-                                     label=< <B>{node_id}</B> >];')
-                dot_lines.append(node.true_pipeline.to_dot(None, _prefix=full_id + "_T_"))
-                dot_lines.append(node.false_pipeline.to_dot(None, _prefix=full_id + "_F_"))
-                true_first = node.true_pipeline.static_order()[0]
-                false_first = node.false_pipeline.static_order()[0]
-                true_last = node.true_pipeline.static_order()[-1]
-                false_last = node.false_pipeline.static_order()[-1]
-                dot_lines.append(f'    "{full_id}" -> "{full_id}_T_{true_first}" '+
-                                 '[label="True", tailport=s];')
-                dot_lines.append(f'    "{full_id}" -> "{full_id}_F_{false_first}" '+
-                                 '[label="False", tailport=s];')
-                dot_lines.append(f'    "{full_id}_output" [shape=diamond, '+
-                                 'label=< <FONT POINT-SIZE="10"> If Output</FONT> >];')
-                dot_lines.append(f'    "{full_id}_T_{true_last}" -> "{full_id}_output" '+
-                                 '[tailport=s];')
-                dot_lines.append(f'    "{full_id}_F_{false_last}" -> "{full_id}_output" '+
-                                 '[tailport=s];')
-                dot_lines.append('  }')
-            elif isinstance(node, NodeFor):
-                dot_lines.append(f'  subgraph cluster_{full_id} {{')
-                dot_lines.append('    style=dashed;')
-                dot_lines.append(f'    "{full_id}" [shape=Mdiamond, label=< <B>{node_id}</B><BR/>'+
-                                 '<FONT POINT-SIZE="10">For Loop</FONT> >];')
-                dot_lines.append(node.loop_pipeline.to_dot(None, _prefix=full_id + "_L_"))
-                loop_first = node.loop_pipeline.static_order()[0]
-                loop_last = node.loop_pipeline.static_order()[-1]
[…]
