+++ src/pipeoptz/__init__.py
@@ -5,8 +5,9 @@ from .parameter import IntParameter, FloatParameter, \
                        BoolParameter, ChoiceParameter, MultiChoiceParameter
 from .pipeline import Pipeline
 from .optimizer import PipelineOptimizer
+from .visualization import Visualizer
 
 __version__ = "0.1.5"
 __all__ = ["Node", "NodeIf", "NodeFor", "NodeWhile", "IntParameter",
            "FloatParameter", "BoolParameter", "ChoiceParameter",
-           "MultiChoiceParameter", "Pipeline", "PipelineOptimizer"]
+           "MultiChoiceParameter", "Pipeline", "PipelineOptimizer", "Visualizer"]
