### 2026-04-05 1388e1c
Adding tests for visualization
