### 2026-04-07 ea5ccea
Adding support for mermaid graphs #4
