+++ src/pipeoptz/visualization.py
@@ -1,7 +1,8 @@
 """Visualization module for Pipeline graphs using DOT/Graphviz."""
 from __future__ import annotations
 import os
-from typing import TYPE_CHECKING, List, Optional
+import re
+from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
 if TYPE_CHECKING:
     from .pipeline import Pipeline
@@ -162,6 +163,205 @@ class Visualizer:
             with open(filepath, "w", encoding="utf-8") as f:
                 f.write(dot_str)
         return "\n".join(dot_lines)
+
+    @staticmethod
+    def _mermaid_id(node_id: str, prefix: str = "") -> str:
+        """Builds a Mermaid-safe identifier."""
+        safe_id = re.sub(r"[^0-9A-Za-z_]", "_", f"{prefix}{node_id}")
+        if safe_id == "":
+            return "node"
+        if safe_id[0].isdigit():
+            return f"n_{safe_id}"
+        return safe_id
+
+    @staticmethod
+    def _mermaid_text(text: str) -> str:
+        """Escapes label text for Mermaid."""
+        return text.replace('"', "'").replace("<lambda>", "lambda")
+
+    @staticmethod
+    def _function_label(node: Node) -> str:
+        """Returns a readable function label for a node."""
+        func_module = node.func.__module__
+        func_name = node.func.__name__
+        if func_name == "<lambda>":
+            return "lambda"
+        if func_module == "__main__":
+            return func_name
+        return f"{func_module}.{func_name}"
+
+    @classmethod
+    def _add_mermaid_node(cls, state: Dict[str, Any], node_id: str,
+                          label: str, shape: str = "box") -> None:
+        """Adds a Mermaid node definition once."""
+        if node_id in state["defined_nodes"]:
+            return
+        wrappers = {
+            "box": ('["', '"]'),
+            "diamond": ('{"', '"}'),
+            "terminal": ('(["', '"])'),
+        }
+        prefix, suffix = wrappers[shape]
+        state["lines"].append(f'    {node_id}{prefix}{cls._mermaid_text(label)}{suffix}')
+        state["defined_nodes"].add(node_id)
+
+    @staticmethod
+    def _add_mermaid_edge(state: Dict[str, Any], source: str,
+                          target: str, label: Optional[str] = None) -> None:
+        """Adds a Mermaid edge definition once."""
+        edge = f"{source}|{label}|{target}" if label is not None else f"{source}|{target}"
+        if edge in state["defined_edges"]:
+            return
+        if label is None:
+            state["lines"].append(f"    {source} --> {target}")
+        else:
+            state["lines"].append(f'    {source} -->|{label}| {target}')
+        state["defined_edges"].add(edge)
+
+    def to_mermaid(self, filepath: Optional[str] = None,
+                   add_optz: bool = False, show_function: bool = True,
+                   _prefix: str = "", _state: Optional[Dict[str, Any]] = None) -> str:
+        """
+        Generates a Mermaid flowchart representation of the pipeline graph.
+
+        Args:
+            filepath (str, optional): The path to save the Mermaid text file.
+                If None, no file is saved.
+
+        Returns:
+            the Mermaid flowchart string of the pipeline
+        """
+        is_root = _state is None
+        if _state is None:
+            _state = {
+                "lines": ["flowchart TD"],
+                "defined_nodes": set(),
+                "defined_edges": set(),
+            }
+
+        last_node_id = self.pipeline.static_order()[-1] if self.pipeline.static_order() else None
+
+        for node_id, node in self.pipeline.nodes.items():
+            full_id = self._mermaid_id(node_id, _prefix)
+            is_last = node_id == last_node_id and _prefix == ""
+
[…]
