+++ src/pipeoptz/pipeline.py
@@ -2,7 +2,6 @@
 from __future__ import annotations
 import json
 import importlib
-import os
 import sys
 import time
 from random import randrange, shuffle
