### 2026-04-08 af5f15a
Bump version to 0.1.6
