### 2026-04-28 #9871 @danielskatz
@erik-whiting  - As track chair, I try to check on submissions in the CSISM track semi-regularly. For this one, what's its status? I can't see that anything has happened in the 3 weeks.  (Also, I think you mistyped @romsson above, so perhaps there wasn't a notification sent)
