### 2026-05-05 3fb0e83
Check instance before using method
