### 2026-05-05 b2c3eee
Replace os.system with subprocess
Co-authored-by: Copilot <copilot@github.com>
