### 2026-05-06 3d9aa41
Write github PR & issues templates
