+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,31 +1,0 @@
----
-name: Bug report
-about: Report a problem or regression
-labels: bug
----
-
-## Description
-
-A clear and concise description of the bug.
-
-## Steps to Reproduce
-
-1. 
-2. 
-3. 
-
-## Expected Behavior
-
-What you expected to happen.
-
-## Environment
-
-- OS: 
-- Python: 
-- Package version/commit: 
-
-## Logs / Traceback
-
-```
-Paste logs here
-```
