+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,21 +1,0 @@
----
-name: Feature request
-about: Suggest an idea or enhancement
-labels: enhancement
----
-
-## Problem to Solve
-
-What problem are you trying to solve?
-
-## Proposed Solution
-
-Describe the change you would like to see.
-
-## Alternatives Considered
-
-What other approaches did you consider?
-
-## Additional Context
-
-Any other context, screenshots, or examples.
