+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,8 +1,5 @@
-blank_issues_enabled: true
+blank_issues_enabled: false
 contact_links:
-  - name: Documentation
-    url: https://ntt.readthedocs.io/
-    about: Check the docs for usage and installation help.
-  - name: GitHub Discussions
-    url: https://github.com/centralelyon/ntt/discussions
-    about: Ask questions or share ideas with the community.
+  - name: Questions and support
+    url: https://github.com/centralelyon/pipeoptz/discussions
+    about: Please use Discussions for questions and support.
