+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,5 @@
 blank_issues_enabled: false
 contact_links:
   - name: Questions and support
-    url: https://github.com/centralelyon/pipeoptz/discussions
-    about: Please use Discussions for questions and support.
+    url: https://github.com/centralelyon/pipeoptz/issues
+    about: Please use Issues for questions and support.
