### 2026-05-11 5defbd8
add Dockerfile
Co-authored-by: Copilot <copilot@github.com>
