### 2026-05-11 b3cc779
add docker tuto in README
