### 2026-07-01 47dc97b
Clarifications and fixes in README.md
