### 2026-07-01 ff5e291
Adding tests in case graphviz is not available
