+++ src/pipeoptz/optimizer.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import random as rd
 import warnings
-from typing import Any, Callable, Dict, Optional, List, Tuple
+from typing import Any, Callable, Dict, Iterable, Optional, List, Tuple, Union
 
 import numpy as np
 from scipy.stats import norm
@@ -15,6 +15,7 @@ from sklearn.exceptions import ConvergenceWarning
 from .parameter import IntParameter, FloatParameter, ChoiceParameter, \
                     BoolParameter, MultiChoiceParameter, Parameter
 from .pipeline import Pipeline, _product
+from .callback import Callback, _CallbackList
 
 warnings.filterwarnings("ignore", category=ConvergenceWarning)
 
@@ -57,6 +58,40 @@ class PipelineOptimizer:
         self.max_time_pipeline: float = max_time_pipeline
         self.loss: Callable[..., float] = loss_function
         self.best_params_history: List[Dict[str, Any]] = []
+        self._callbacks = _CallbackList()
+        self._evaluation_count = 0
+
+    def _on_iteration_begin(
+        self, method: str, iteration: int, total: int, unit: str = "iteration"
+    ) -> None:
+        """Notify callbacks that an optimization iteration is starting."""
+        self._callbacks.call(
+            "on_iteration_begin",
+            iteration,
+            logs={"method": method, "total_iterations": total, "unit": unit},
+        )
+
+    def _on_iteration_end(
+        self,
+        method: str,
+        iteration: int,
+        total: int,
+        best_loss: float,
+        best_params: Dict[str, Any],
+        unit: str = "iteration",
+    ) -> None:
+        """Notify callbacks that an optimization iteration has finished."""
+        self._callbacks.call(
+            "on_iteration_end",
+            iteration,
+            logs={
+                "method": method,
+                "total_iterations": total,
+                "unit": unit,
+                "best_loss": float(best_loss),
+                "best_params": best_params.copy(),
+            },
+        )
 
     def add_param(self, param: Parameter) -> None:
         """Adds a parameter to the list of parameters that the optimizer will tune."""
@@ -128,6 +163,16 @@ class PipelineOptimizer:
             y_negative (list, optional): A list of negative examples for some loss functions. 
                 Defaults to None.
         """
+        evaluation = self._evaluation_count
+        self._evaluation_count += 1
+        evaluation_logs = {
+            "params": self.get_params_value().copy(),
+            "sample_count": len(X),
+        }
+        self._callbacks.call(
+            "on_evaluation_begin", evaluation, logs=evaluation_logs.copy()
+        )
+
         self.update_pipeline_params()
         is_triplet_loss = y_negative is not None
 
@@ -142,8 +187,20 @@ class PipelineOptimizer:
             else:
                 loss += self.loss(results[-1], y[i])
             if self.max_time_pipeline != 0 and t[0] > self.max_time_pipeline:
-                return results+[None]*(len(X)-i-1), float("inf")
+                results += [None] * (len(X) - i - 1)
+                loss = float("inf")
+                self._callbacks.call(
+                    "on_evaluation_end",
+                    evaluation,
+                    logs={**evaluation_logs, "loss": loss, "timed_out": True},
+                )
+                return results, loss
         loss /= len(X)
+        self._callbacks.call(
+            "on_evaluation_end",
+            evaluation,
+            logs={**evaluation_logs, "loss": float(loss), "timed_out": False},
+        )
         return results, loss
 
     def plot_convergence(self) -> np.ndarray:
@@ -260,6 +317,7 @@ class PipelineOptimizer:
         loss_log = []
 
         for i in range(iterations):
[…]
