+++ src/pipeoptz/__init__.py
@@ -5,9 +5,11 @@ from .parameter import IntParameter, FloatParameter, \
                        BoolParameter, ChoiceParameter, MultiChoiceParameter
 from .pipeline import Pipeline
 from .optimizer import PipelineOptimizer
+from .callback import Callback
 from .visualization import Visualizer
 
 __version__ = "0.1.6"
 __all__ = ["Node", "NodeIf", "NodeFor", "NodeWhile", "IntParameter",
            "FloatParameter", "BoolParameter", "ChoiceParameter",
-           "MultiChoiceParameter", "Pipeline", "PipelineOptimizer", "Visualizer"]
+           "MultiChoiceParameter", "Pipeline", "PipelineOptimizer", "Callback",
+           "Visualizer"]

[…]
