### 2026-07-02 c4fa55f
Clarifying support and install for MkDocs Material
