### 2026-08-21 #9871 @erik-whiting
> 👋 - [@erik-whiting](https://github.com/erik-whiting) - can you provide a short update on how this review is going?

The reviewer @jqwang2373 stated that they are re-reviewing the submission. This was two weeks ago. I pinged the other reviewers 3 weeks ago to continue review but have not heard back. Given that the most recent update was during the end of the summer, I am assuming reviewers are busy with the start of the Fall semester and did not want to poke them again this early.

This review started in January (start of Spring semester) and the reviewers did not submit their first reviews until a month after. I am counting on the second round of reviews to follow the same pattern; I'm assuming that if I poke reviewers now, the notification will be buried in things more pressing to the reviewers. Planning to ping them again in the second week of September when the initial rush of the semester has passed.
