### 2026-08-30 #9871 @erik-whiting
@gflofst can you take a glance at this paper and see if this original concern of yours is adequately addressed:

>There are a TON of engines that could be referenced, but are not. It is not necessary to talk about all of them as that would be a long, detailed survey paper. However, it would be good to scope it down so that the Pegasus/Condor combo that does the same things as this software, but in a vastly richer tool (and vastly more complex) is mentioned, but sidelined for reasonable reasons.

From @shrishar's review, it looks like the sufficiency of this software's novelty for publication is in question.

@romsson can you confirm if you've seen the latest review comment? I don't need a comprehensive response, but could you let us know if you have any general thoughts on this specific statements:

>I am not fully convinced about its novelty relative to existing workflow and optimization frameworks such as Dask, FireWorks, Pegasus, and Bayesian optimization libraries.

and

>The Research Impact Statement is also somewhat weak: the authors mainly cite its use in an ongoing centralelyon/ntt prototype, PyPI availability, documentation, testing, and examples. These demonstrate software maturity and reusability, but provide limited evidence of research impact

Let me know, thanks!
